+++ src/main/java/urlshortener/service/systeminformation/SystemCalculatorService.java
@@ -1,15 +1,15 @@
-package urlshortener.information;
+package urlshortener.service.systeminformation;
 
 import org.springframework.stereotype.Component;
 import urlshortener.service.ClickService;
 import urlshortener.service.ShortURLService;
 
 @Component
-public class SystemCalculator {
+public class SystemCalculatorService {
   private final ClickService clickService;
   private final ShortURLService shortUrlService;
 
-  public SystemCalculator(ClickService clickService, ShortURLService shortUrlService) {
+  public SystemCalculatorService(ClickService clickService, ShortURLService shortUrlService) {
     this.clickService = clickService;
     this.shortUrlService = shortUrlService;
   }
